+++ CHANGELOG.md
@@ -1,0 +1,12 @@
+### v1.0.2
+* Core is now downloading dynamic modules by itself, without bindings outlets;
+* fixed premature after hook trigger when dynamic component is downloading;
+* initializers for router modes are now in separated modules;
+* reorganized file system.
+
+### v1.0.1
+* Separated file for utils methods for better module system.
+
+### v1.0.0
+* Launched separated module;
+* component type inherits any for different libs compatibility.
